Make queue work a bit more like other music players
When you start playing a song, it will automatically replace all items
in the queue. At some point we should maybe add a dialog if you have a
lot of items in the queue that you will be overwriting, but this way
the queue is always synced up to what's playing.

This also refactors the queue and player so that to play music you
always have to go through the queue.